Check event details for updates from the server
Event details keep changing. This commit checks if they have changed. As each check takes some time, the user is shown cached data while it checks for updates in the background.
It only checks for cached data if:
1) It has been over 5 minuites since the last check
2) The event details version has changed from that known to the current Event

This commit shows some data to the user, even if it is not current, and reduces the number of times the app has to make requests to the server, because that is limited. The downside is that it will take several seconds to load the new data, in which time the user will have no idea that they are reading stale data.